Try to Signin silently while access token expired.
- Before show access token expired dialog try to renew the access token few time
if it was not succeed then show access token expired dialog and when user accept that dialog
redirect to logon page (prompt=select_account).
- Handle the oidc events which used renew access token and on fails it will
redirect to interactive login page.
- Before redirect to logon page clear the php session.
- Refine the wrapperHandler function it will check if Extjs is available then
it will show the Extjs message box else it will show the custom message box.
- Retry button of message box should retry to validate the token.
- Redirect to interactive login page when user clicking on 'yes' button
of session expired dialog while clicking on 'No' button it will remains in
webapp.
- Add 'reauthenticateWithRedirect' function which redirect to interactive login page.
Function triggered when 'yes' button of session expired dialog was pressed.
- Try to renew an access token once when silent renew error event triggered
 and user was expired already. If it fails show access token expired message box.
- Add logs in oidc-kopano.js to make debugging easy.

Reference: KW-3163

+++ server/includes/controllers/service.fingerprint.php
@@ -57,7 +57,7 @@
 	
 	// Single sign on will never go through the login page. So we cannot check
 	// the fingerprint there!
-	if ( !WebAppAuthentication::isUsingSingleSignOn() && (!isset($_SESSION['frontend-fingerprint']) || $_POST['fingerprint'] !== $_SESSION['frontend-fingerprint']) ){
+	if ( empty(OIDC_ISS) && !WebAppAuthentication::isUsingSingleSignOn() && (!isset($_SESSION['frontend-fingerprint']) || $_POST['fingerprint'] !== $_SESSION['frontend-fingerprint']) ){
 		error_log('frontend-fingerprint did not match. Session terminated. ' . WebAppAuthentication::getUserName());
 		$phpSession->destroy();
 		Response::unAuthorized();
